+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed validation of FEP-1b12 activities where `object.actor` is embedded.
+- Prevent removal of attachments during post editing in some Mastodon API clients.
 
 ### Security